Organized components and MainComponent added
