+++ simple-table/src/components/table/ProductTable.js
@@ -3,52 +3,51 @@ import ProductRow from './ProductRow';
 import ProductCategoryRow from './ProductCategoryRow';
 
 class ProductTable extends Component {
-    render() {
+  render() {
+    const filterText = this.props.filterText;
+    const inStockOnly = this.props.inStockOnly;
 
-      const filterText = this.props.filterText;
-      const inStockOnly = this.props.inStockOnly;
-
-      const rows = [];
-      let lastCategory = null;
+    const rows = [];
+    let lastCategory = null;
   
-      this.props.products.forEach((product) => {
-        if(product.name.indexOf(filterText) === -1) {
-            return;
-        }
-        if(inStockOnly && !product.stocked) {
-            return;
-        }
-        if(product.category !== lastCategory) {
-          rows.push(
-            <ProductCategoryRow
-              category = {product.category}
-              key = {product.category}
-            />
-          );
-        }
+    this.props.products.forEach((product) => {
+      if(product.name.indexOf(filterText) === -1) {
+        return;
+      }
+      if(inStockOnly && !product.stocked) {
+        return;
+      }
+      if(product.category !== lastCategory) {
         rows.push(
-          <ProductRow
-            product = {product}
-            key = {product.name}
+          <ProductCategoryRow
+            category = {product.category}
+            key = {product.category}
           />
         );
-        lastCategory = product.category;      
-      });
-  
-      return(
-        <table>
-          <thread>
-            <tr>
-              <th>Name</th>
-              <th>Price</th>
-            </tr>
-          </thread>
-          <tbody>
-            {rows}
-          </tbody>
-        </table>
+      }
+      rows.push(
+        <ProductRow
+          product = {product}
+          key = {product.name}
+        />
       );
-    }
+      lastCategory = product.category;      
+    });
+  
+    return(
+      <table>
+        <thread>
+          <tr>
+            <th>Name</th>
+            <th>Price</th>
+          </tr>
+        </thread>
+        <tbody>
+          {rows}
+        </tbody>
+      </table>
+    );
+  }
 }
   
 export default ProductTable;
